+++ src/femagtools/machine/afpm.py
@@ -4,13 +4,16 @@
 import logging
 import numpy as np
 from pathlib import Path
+import shutil
 from .. import poc
 from .. import parstudy
 from .. import model
 from .. import utils
 from .. import windings
-from scipy.interpolate import RegularGridInterpolator, interp1d
+from .. import femag
+from scipy.interpolate import RegularGridInterpolator, interp1d, RectBivariateSpline
 from scipy.integrate import quad
+import copy
 
 logger = logging.getLogger(__name__)
 
@@ -36,6 +39,50 @@ def _integrate1d(radius, val):
         return interp((x))
     return quad(func, radius[0], radius[-1])[0]
 
+def ld_interpol(i1, beta, v):
+    '''interpolate Ld at beta angle 0°, -180°'''
+    # ld
+    cur = copy.deepcopy(i1)
+    betad = copy.deepcopy(beta)
+    if np.amin(beta) < -90 and \
+        np.amax(beta) > -90: 
+        # motor and generator
+        v[0] = v[1]
+        v[-1] = v[-2]
+
+        dbeta = np.abs(beta[0][0] - beta[1][0])
+        bp = [[beta[0][0]-dbeta for i in range(len(np.unique(i1)))]] + beta[1:-1] + \
+            [[dbeta for i in range(len(np.unique(i1)))]]
+    else: 
+        v[-1] = v[-2]
+        dbeta = np.abs(beta[0][0] - beta[1][0])
+        bp = beta[0:-1] + \
+            [[dbeta for i in range(len(np.unique(i1)))]]
+    
+    return RectBivariateSpline(np.unique(bp), np.unique(cur), \
+         np.array(v)).ev(*[betad, i1]).tolist()
+
+def lq_interpol(i1, beta, v):
+    '''interpolate Lq at beta -90°'''
+    if -90 not in np.unique(beta): 
+        return v
+    # lq
+    betad = copy.deepcopy(beta)
+    if np.amin(beta) < -90 and \
+        np.amax(beta) > -90:
+        # motor and generator
+        inx = np.argwhere(np.array(beta) == -90).squeeze()
+        v.pop(inx[0, 0])
+        bp = beta[0:inx[0, 0]] + beta[inx[0, 0]+1:]
+        cp = i1[0:inx[0, 0]] + i1[inx[0, 0]+1:]
+    else: 
+        v[0] = v[1]
+        dbeta = np.abs(beta[0][0] - beta[1][0])
+        bp = [[-90-dbeta for i in i1[0]]] + beta[1::]
+        cp = i1
+    cur = copy.deepcopy(cp)
+    return RectBivariateSpline(np.unique(bp), np.unique(cur), \
+         np.array(v)).ev(*[betad, i1]).tolist()
 
 def parident(workdir, engine, temp, machine,
              magnetizingCurves, magnetMat=[], condMat=[],
@@ -124,10 +171,6 @@ def parident(workdir, engine, temp, machine,
         ]
     }
 
-    pstudy = parstudy.List(
-        workdir, condMat=condMat, magnets=magnetMat,
-        magnetizingCurves=magnetizingCurves,
-        cmd=kwargs.get('cmd', None))
 
     ldq = []
     for magtemp in temp:
@@ -138,12 +181,37 @@ def parident(workdir, engine, temp, machine,
             poc=poc.Poc(999),
             speed=0)
         logging.info("Noload simulation")
-        nlresults = pstudy(nlparvardef, machine, nlcalc, engine)
-        if nlresults['status'].count('C') != len(nlresults['status']):
-            raise ValueError('Noload simulation failed %s', nlresults['status'])
-        nlresults.update(process(lfe, pole_width, machine, nlresults['f']))
-        current_angles = nlresults['f'][0]['current_angles']
+        if kwargs.get('use_multiprocessing', True): 
+            pstudy = parstudy.List(
+                    workdir, condMat=condMat, magnets=magnetMat,
+                    magnetizingCurves=magnetizingCurves,
+                    cmd=kwargs.get('cmd', None))
 
+            nlresults = pstudy(nlparvardef, machine, nlcalc, engine)
+            if nlresults['status'].count('C') != len(nlresults['status']):
+                raise ValueError('Noload simulation failed %s', nlresults['status'])
+        else: 
+            nlresults = {"x": [], "f": []}
+            i = 0
+            for pw, le, sp in zip(pole_width, lfe, linspeed): 
+                nlmachine = {k: machine[k] for k in machine}
+                nlmachine['pole_width'] = pw
+                nlmachine['lfe'] = le
+                nlcalc.update({"speed": sp})
+                nlsubdir = f'{workdir}/{i}'
+                nlworkdir = Path(nlsubdir)
+                if nlworkdir.exists(): 
+                    shutil.rmtree(nlworkdir)
+                nlworkdir.mkdir(exist_ok=True)
+                noloadsim = femag.Femag(nlworkdir, condMat=condMat, magnets=magnetMat,
+                                        magnetizingCurves=magnetizingCurves,
+                                        cmd=kwargs.get('cmd', None))
+                r = noloadsim(nlmachine, nlcalc)
+                nlresults['f'].append({k: v for k, v in r.items()})
+                i = i + 1
+        nlresults.update(process(lfe, pole_width, machine, nlresults['f']))
+
+        current_angles = nlresults['f'][0]['current_angles']
         results = []
         i = 0
         for l, pw in zip(lfe, pole_width):
@@ -152,17 +220,11 @@ def parident(workdir, engine, temp, machine,
             mpart['lfe'] = l
             subdir = f"{workdir}/{i}"
 
-            gpstudy = parstudy.Grid(
-                subdir, condMat=condMat, magnets=magnetMat,
-                magnetizingCurves=magnetizingCurves,
-                cmd=kwargs.get('cmd', None))
-
             simulation = dict(
                 calculationMode="torq_calc",
                 wind_temp=20.0,
                 magn_temp=magtemp,
                 angl_i_up=0.0,
-                magnet_loss=True,
                 magn_height=machine['magnet']['afm_rotor']['magn_height'],
                 yoke_height=machine['magnet']['afm_rotor'].get(
                     'yoke_height', 0),
@@ -173,8 +235,36 @@ def parident(workdir, engine, temp, machine,
                 num_move_steps=60,
                 speed=linspeed[i],
                 num_par_wdgs=machine[wdgk].get('num_par_wdgs', 1))
+            
+            if kwargs.get('use_multiprocessing', True): 
+                gpstudy = parstudy.Grid(
+                                    subdir, condMat=condMat, magnets=magnetMat,
+                                    magnetizingCurves=magnetizingCurves,
+                                    cmd=kwargs.get('cmd', None))
+                lresults = gpstudy(parvardef, mpart, simulation, engine)
 
-            lresults = gpstudy(parvardef, mpart, simulation, engine)
+            else:
+                lresults = {"x": [], "f": []}
+                domain_beta = np.linspace(beta_min, 0, num_beta_steps).tolist()
+                domain_cur = np.linspace(i1_max/num_cur_steps, i1_max, num_cur_steps).tolist()
+                dir_index = 0
+                for cur in domain_cur: 
+                    for be in domain_beta:
+                        simulation['angl_i_up'] = be
+                        simulation['current'] = cur
+                        lresults['x'].append([be, cur])
+                        subsubdir = subdir + f'/{dir_index}'
+                        dir_index = dir_index + 1
+                        lworkdir = Path(subsubdir)
+                        if lworkdir.exists(): 
+                            shutil.rmtree(lworkdir)
+                        lworkdir.mkdir(exist_ok=True)
+                        loadsim = femag.Femag(lworkdir, condMat=condMat, magnets=magnetMat,
+                                            magnetizingCurves=magnetizingCurves,
+                                            cmd=kwargs.get('cmd', None))
+                        r = loadsim(mpart, simulation)
+                        lresults['f'].append({k: v for k, v in r.items()})
+                        
             f = [{k: bch[k]
                   for k in ('linearForce', 'flux', 'losses', 'lossPar')}
                  for bch in lresults['f']]
@@ -198,6 +288,29 @@ def parident(workdir, engine, temp, machine,
                           (-1, num_beta_steps)).T/np.sqrt(2)
         psiq = np.reshape([r['psiq'] for r in postp],
                           (-1, num_beta_steps)).T/np.sqrt(2)
+        
+        ld = np.reshape([r['Ld'] for r in postp],
+                          (-1, num_beta_steps)).T.tolist()
+        lq = np.reshape([r['Lq'] for r in postp],
+                          (-1, num_beta_steps)).T.tolist()
+        # interpolation ld, lq
+        curr, angl = [], [] 
+        for cr in range(len(beta)): 
+            curr.append(i1)
+        for al in beta: 
+            tmp = []
+            for cr in range(len(i1)): 
+                tmp.append(al)
+            angl.append(tmp)
+        try:
+            xx, yy = copy.deepcopy(curr), copy.deepcopy(angl)
+            ld = ld_interpol(xx, yy, ld)
+            xx, yy = copy.deepcopy(curr), copy.deepcopy(angl)
+            lq = lq_interpol(xx, yy, lq)
+        except: 
+            ld = np.zeros_like(psid).tolist()
+            lq = np.zeros_like(psid).tolist()
+        
         torque = np.reshape([r['torque'] for r in postp],
                             (-1, num_beta_steps)).T
         losses = {k: np.flip(np.reshape([r['plfe'][k] for r in postp],
@@ -214,6 +327,7 @@ def parident(workdir, engine, temp, machine,
         ldq.append({'temperature': magtemp,
                     'i1':i1, 'beta':beta,
                     'psid': psid.tolist(), 'psiq': psiq.tolist(),
+                    'ld': ld, 'lq': lq,
                     'torque': torque.tolist(),
                     'losses': losses})
         # T = 3/2 p (Psid iq - Psiq id)