+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 	This document lists the major changes in femagtools. Please clone this project to follow bug fixes and minor enhancements.
 
+## Release 1.5.5
+	- PM, SM characteristics: auto correction of max torque
+
+## Release 1.5.2
+	- VTU demag access by elements added
+
 ## Release 1.4.8
 	- Core loss calc method Bertotti added